Add basic JavaFX UI, move old code to deprecated folder
