+++ src/deprecated/GraycodeReflectedBinaryMappingSOP.java
