+++ src/deprecated/KarnaughMapRenderer.java
