+++ src/deprecated/QuineMcCluskey.java
@@ -1,0 +1,10 @@
+/**
+ * Created by horatiulazu on 2017-06-05.
+ */
+public class QuineMcCluskey {
+
+
+    public QuineMcCluskey(){
+
+    }
+}
