+++ src/deprecated/TruthTable.java
