+++ src/deprecated/tests/CompoundedGroupings.java
