+++ src/deprecated/tests/CoordinateTest.java
