+++ src/deprecated/tests/FourVariableTest.java
