+++ src/deprecated/tests/SimplificationTest.java
