+++ src/deprecated/tests/ThreeVariableTest.java
