+++ src/deprecated/tests/TruthTableTest.java
