+++ src/main/java/Controller.java
@@ -1,0 +1,5 @@
+/**
+ * Created by horatiulazu on 2017-06-17.
+ */
+public class Controller {
+}
