+++ src/test/java/Test.java
@@ -1,0 +1,12 @@
+import junit.framework.TestCase;
+
+/**
+ * @author Horatiu Lazu
+ * @version 1.0
+ */
+public class Test extends TestCase{
+    public void testTest() throws Exception {
+        assertEquals(true, true);
+    }
+
+}
